+++ lib/Extension/ExpressionExtension.php
@@ -44,7 +44,7 @@ use PhpBench\Expression\NodeEvaluator\DisplayAsEvaluator;
 use PhpBench\Expression\NodeEvaluator\FunctionEvaluator;
 use PhpBench\Expression\NodeEvaluator\ListEvaluator;
 use PhpBench\Expression\NodeEvaluator\LogicalOperatorEvaluator;
-use PhpBench\Expression\NodeEvaluator\ParameterEvaluator;
+use PhpBench\Expression\NodeEvaluator\PropertyAccessEvaluator;
 use PhpBench\Expression\NodeEvaluator\ParenthesisEvaluator;
 use PhpBench\Expression\NodeEvaluator\PhpValueEvaluator;
 use PhpBench\Expression\NodeEvaluator\TolerableEvaluator;
@@ -188,7 +188,7 @@ class ExpressionExtension implements ExtensionInterface
                 new ComparisonEvaluator(),
                 new TolerableEvaluator(),
                 new DisplayAsEvaluator(),
-                new ParameterEvaluator(),
+                new PropertyAccessEvaluator(),
                 new ConcatEvaluator(),
                 new PhpValueEvaluator(),
             ]);
